refactor: 🛠 fetching news at build time
