+++ components/common/NewsAppearance.tsx
@@ -10,12 +10,12 @@ import {
   useBreakpointValue,
   VStack,
 } from "@chakra-ui/react";
-import { useNewsList } from "@shm/hooks";
 import Image from "next/image";
 import Link from "next/link";
 import { Swiper, SwiperSlide, useSwiper } from "swiper/react";
 import { IconLeftArrow, IconRightArrow } from "./Icons";
 import SectionHeading from "./SectionHeading";
+import { NewsItem } from "../../types";
 
 const SliderButtons = () => {
   const swiper = useSwiper();
@@ -73,9 +73,8 @@ const NewsItem = ({
   </Link>
 );
 
-const NewsAppearance = ({ sectionTitle }: { sectionTitle: string }) => {
+const NewsAppearance = ({ sectionTitle, news }: { sectionTitle: string; news: NewsItem[] }) => {
   const slidesPerView = useBreakpointValue({ base: 1, md: 2 });
-  const { data } = useNewsList();
 
   return (
     <Box>
@@ -94,7 +93,7 @@ const NewsAppearance = ({ sectionTitle }: { sectionTitle: string }) => {
             </HStack>
           </VStack>
         </Box>
-        {data.map((item) => (
+        {news.map((item) => (
           <SwiperSlide key={item.title}>
             <NewsItem
               title={item.title}
