+++ components/sections/home/ShardeumInNews.tsx
@@ -3,8 +3,9 @@ import React from "react";
 import JoinNewsletter from "components/common/JoinNewsletter";
 import NewsAppearance from "components/common/NewsAppearance";
 import VisitBlog from "components/common/VisitBlog";
+import { NewsItem } from "../../../types";
 
-function ShardeumInNews() {
+function ShardeumInNews({ news }: { news: NewsItem[] }) {
   return (
     <Flex as="section" bg="brand.grey-10">
       <Container
@@ -14,7 +15,7 @@ function ShardeumInNews() {
         px={{ base: "6", xl: 0 }}
       >
         <Box mb={{ base: "10", md: "20", lg: "32" }}>
-          <NewsAppearance sectionTitle="Shardeum In News" />
+          <NewsAppearance sectionTitle="Shardeum In News" news={news} />
         </Box>
         <SimpleGrid columns={{ base: 1, lg: 2 }} spacing="6" w="full">
           <JoinNewsletter />
